cache proxy: dont include cached digest key when hashing GetActionResultRequest (#8832)
Previously, if we requested a proxy-cached ActionResult and the remote
server has a _different_ ActionResult, we would use a modified
GetActionResultRequest that included cached_action_result_digest as a
key when saving the ActionResult to the cacheproxy cache. This meant
that the cacheproxy-cached ActionResult was never getting overwritten!

This change computes the AC key for the request prior to setting
cached_action_result_digest.