Respect min/max duration filters in invocation aggregate/summary queries (#8807)
Was trying to filter the UI to show only prober invocations > 1 minute
and noticed the invocation list was being updated but not the summary at
the top which shows the total invocation count / duration etc. This PR
fixes that.

Also simplify some of the server-side query logic and fix some TS type
issues (we were using the `InvocationQuery` message for the
`GetInvocationStatRequest.query` field but should be using
`InvocationStatQuery`).