SCIM: fix user creation if user exists but not in SCIM group
From the perspective of an external system, a user does not exist if
it's not a member of the SCIM-managed group. We do not return such users
via the query APIs.

If a user is manually removed via the UI, but is still mapped in the
external system then the external system will send a "create user"
request to bring the systems in sync. Prior to this fix, this would fail
since the internal create call would fail due to the existence of a user
with the same SubID.