[FC] Add experiment to skip resaving snapshots (#8830)
This is all the code needed to trigger the experiment based on metadata
or executor hostname, and to log the experiment to clickhouse.

https://github.com/buildbuddy-io/buildbuddy/pull/8808 implements the
firecracker changes.

https://github.com/buildbuddy-io/buildbuddy-internal/pull/4717 is an
example for configuring the experiment.

Triggering the experiment by either request metadata or executor
hostname makes it more complicated.

Each task is enqueued on multiple executors, and we can't add the
platform property to the task until we know which executor will dequeue
it, because we only have 1 copy of the task on the scheduler side. This
means we have to wait until Scheduler.Lease to set the platform property
to triggering on executor hostname.

The executor doesn't have the task (and therefore the platform
properties) until Lease returns. It also doesn't have Bazel's
RequestMetadata until then, which means Scheduler.Lease doesn't get the
metadata either, so we need some hacks to add the metadata values to the
experiment context.