cli/explain: replace string interner with unique.Make
This replace our manual string interner with Go's new 'unique' package
that does the same thing. https://go.dev/blog/unique

The main differences here is that the 'unique' package leverages the
new 'weak' pointers to store string values, which should be more
gc-friendly than a growing map[string]string.

A better option here would be to leverage VTProtobuf's
`(vtproto.options).unique` on relevant string fields so that unique.Make
is applied automatically when we unmarshal the proto. Though we shall
leave that for a later date.