.bazelrc: introduce shared.bazelrc (#8828)
Split most of common bazelrc flags into a shared file so that we can
sync and reuse it with our internal repo.

Intentionally exclude most of environment-specific configs as they can
be a bit cummbersome to untangle in this change.

One important change in this PR is the removal of the default flag
`--noremote_upload_local_results`. The reasoning is that uploading AC
should be relatively cheap vs CAS so we don't really have a strong
reason to skip it. The flag is no-op unless a `--remote_cache` or
`--remote_executor` is set, and we do enable it inside `cache-shared`
unconditionally. This should bring our config more in sync with
internal's bazelrc.