go.mod: update package name to include /v2
As we migrate to bzlmod, go_repository dependencies will no longer be
managed from deps.bzl but instead, automatically generated by parsing
go.mod content.

This is fine for our public repo, but our internal repo will not have
all the go dependencies needed to build our internal tools. So we need
to migrate the internal repo to Go Module as the first step to unlock
this.

When attempted to do that inside our internal repo, this would be the
result:

  ```
  > cd <internal-repo>
  > go mod init
  > go mod tidy -e
  > rg github.com/buildbuddy-io/buildbuddy go.mod
  ...
  11:	github.com/buildbuddy-io/buildbuddy v1.9.2
  ```

Note that at the moment of writing, our latest release is v2.146.0.
When attempted to upgrade the go module dep to the latest version, this
error showed up:

  ```
  > go get github.com/buildbuddy-io/buildbuddy@v2.146.0
  go:
    github.com/buildbuddy-io/buildbuddy@v2.146.0:
      invalid version:
        module contains a go.mod file, so module path must match major version ("github.com/buildbuddy-io/buildbuddy/v2")
  exit 1
  ```

So we need a valid go.mod in our public repo to fix this.