Store command_snippet in ClickHouse (#8894)
This is a blocker for MySQL => ClickHouse migration since we currently
show this data in the UI and it's driven by the command_snippet field
that we store in MySQL but not ClickHouse.

Longer-term we discussed using `output_paths[0]` instead of command
snippet for display in the UI, but I think we should decouple that from
the MySQL => ClickHouse migration. If we switch to `output_paths[0]`
later, we can always come back and delete this `command_snippet` column.