Fix for line-wrap bug (#8904)
This "fixes" the annoying bug where bazel progress lines get scrolled up
above the line where curses believes them to be, effectively freezing
those lines in time and then beginning the progress lines below. I
discovered that what was happening was that particulary long lines (>100
characters) would wrap in the ANSI state machine we use, which would
cause any lines above it to move up. That means any lines above that
line would no longer be where curses believed them to be, so the lines
above the window curses modifies would never be changed again, resulting
in the behavior we were seeing.

I increased the number of columns in the state machine from 160 to 256
via the `SetSize` function, which at least makes this much less common.
I am open to suggestions as to what this value ought to be; I selected
256 arbitrarily and then tested it on our repo and saw that the bug did
not occur. The _correct_ solution to this is to decouple the maximum
columns and the cap size of line slices in the `terminal-to-html`
library, which would allow us to use an arbitrarily large number (like
`math.MaxInt`) as the column size, at which point this bug would never
occur again instead of just being less common. However, this is very
low-hanging fruit that fixes the 95% case and does not make the
remaining cases any worse.