Test redacting passwords in URLs in a single function (#8910)
This change groups the tests for redacting passwords in URLs into a
single test function with multiple test cases. It also uses the
`protocmp` and `go-cmp` packages to compare BuildEvents.

In general, I'd like one test function per _kind_ of secret we're
redacting (such as passwords in URLs, non-allowed environment variables,
or repo credentials) with one test case per _place_ where a secret could
appear.

This way, if we find we're failing to redact a particular _kind_ of
secret, we can add or modify the test cases for that _kind_ of secret
and know we've covered it. Same goes for cases where we're
over-redacting.

If people approve of this approach, I'll keep going!