api/common: fix potential panic (#8903)
This PR contains a stack of changes that improve the BES API inside
common.go.

First, we setup a test inside `api_server_test.go` that is based on
actual event data instead of
mock data. This should help us verify our logic against actual Bazel
events.
The existing data is an invocation in dev environemtn with all
--client_env flag removed to avoid
any sensitive information.

The rest of the changes are improvements to common.go to avoid nil
panics issues.