docs: recommend 600 seconds timeout
This has been a recurring theme for a while now that Bazel can wait for
an action for a full hour before issuing a retry.

Let's recommend a moderate timeout value of 10 minutes (600 seconds)
instead. This does mean that larger blob transfers may fail in slow
network conditions. But that's relatively rarer and easier to
troubleshoot.