bzlmod: prepare deps for migration (#8939)
Cleanup MODULE.bazel file to prepare for eventual migration.
The most important change here is that we are explicitly registering our
bb toolchains to override the default auto-registering order, which put
rules_cc local toolchain up top based on the bazel_dep loading order.

In the future we might need to apply the same manual registration to
select the exact Go toolchains that we want to use.

Upgraded googleapis, platforms, rules_python to latest release on
central
registry. This let us get rid of the deprecated switched_rules
extensions call.