Use environment variable to identify if the bb binary is running as the sidecar (#8940)
Switch from searching the os args to try to remove a flag indicating the
sidecar, if it exists, to using an environment variable. This avoids
having to modify os args without parsing them, which is brittle, as well
as parsing the os args, which we would prefer not to do if we are not
the sidecar.