Disable locking in VFS implementation.
There are different types of locks (POSIX/OFD) with different semantics
and the fuse implementation does not have enough information to know
which one is being used (!).

Thankfully if locking is not enabled in the FUSE implementation it is
handled entirely by the kernel which is able to properly handle
different lock types.

I'll rip out the rest of the locking code after I'm satisfied that it
won't be needed anymore.