Allow filtering targets by branch name
We already store branch name in the DB.
This change allows filtering by branch.

Queries with branch filters enabled should be roughly as fast as
what we've got today since we're scanning the same number of rows.
We could speed these up in the future by adding a data-skipping
index on branch name.