[raft] fix TestCleanupZombieInitialMemberNotSetUp (#8971)
This test sometimes failed with the "the shard is not ready".

Upon investigation, this happens with the step in StartShard when we try
to call
sync propose the batch request because we only set up one replica of the
shard
when there are three initial members.

I removed the batch request from the StartShardRequest because this is
not
necessary for the test.

Also add more context in the returned errors in StartShard.