Lock less aggressively in hit-tracker-client (#8955)
I added a small benchmark in the test that enqueues 1,000,000 hits (all
for the same group) asynchronously in batches of 100,000, waiting for
each enqueue() call in a batch to return before proceeding to the next
batch. Before this change, that took 7.26 seconds, with this change it
takes 2.67 seconds (63% faster) with `--config=race --config=remote`.
With `-c opt` it took 3.1 seconds before this change and about 500ms
after. Interestingly, I implemented the channel approach we discussed a
week or so ago and that took about the same amount of time, so there may
be some other resource contention in the benchmark that's limiting
further speedup in the benchmark. For now, this is at least a fair bit
faster than the old approach.