track full response size in action_cache_server even for client-side cache hits. (#8929)
this is admittedly a little weird. there are two options here that i
see:
1. track size in the cache proxy when the value is in the cache proxy's
cache, track size in the apps when it's there instead. this is a little
weird, because we don't know if the value will match until we've
actually read it, at which point we've already initialized a tracker in
the app server, yada yada.
2. do what's in this pr: never track download for ac requests in the
cache proxy. this is also a little weird, but the net result is that we
run no risk of double-tracking hits. i considered this more reasonable
because we literally always perform the reads in the apps under the
current implementation--we _never_ put ac entries in the lookaside
cache, and cut around the hit-tracking service implementations when
looking up values in the cache.