[raft] remove transfer leadership step from removeZombie (#8981)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4192

The current code can sometimes causes one zombie stuck in the transfer
leadership step when we remove zombies that are not in the meta range.
This is because, when we determine the target replica id for leader
transfer, we
are using task.shardInfo.Replicas, which can be stale when we are
removing three
zombies simultaneously on 3 machines.

If we really want to make this work, we probably need to call get
membership to
get the current replicas to find a target. This can make the code more
complicated and I don't think explicit leadership transfer is a
necessary step
anyway.

Also, explicitly disable driver in some zombie test. And set up three
stores instead of 2 to reduce timeout errors.