bzlmod: patch rules_webtesting and force stricter go_dep
rules_webtesting does not have a valid git tag for Go modules, so the
version we are using in go.mod and the actual version would mismatch
each others.

This mismatch would result in several warning in the past, which we have
ignored. However, we do want to elevate this class of issue to error so
that we can catch mismatched Go dependencies earlier.

Patch the bzlmod version of rules_webtesting to be consistent with the
go.mod version. Increase the strictness of go_dep extension to fail when
transitive go deps use a different version than the one in our go.mod.

Ref: https://github.com/bazelbuild/rules_webtesting/issues/452