Allow filtering tests and flakes by branch name (#8970)
We already store branch name in the DB but don't allow filtering on it.
This PR allows us to filter by branch.

Queries with branch filters enabled should be roughly as fast as what
we've got today since we're scanning the same number of rows. We could
speed these up in the future by adding a data-skipping index on branch
name.


https://github.com/user-attachments/assets/14cc4c42-2b6e-42e2-9f8a-0ffaecc71a28