Move cli `Option` and `Definition` into their own file (#8993)
This moves the `Option` and `Definition` struct definitions and
accompanying functions out of the `parser.go` file, which is already
quite large, and into their own file. It also abstracts the `Definition`
somewhat, preventing most access to the definition fields post-creation
and allowing easy access to those accessor functions via the Option.
These changes are in preparation for making the `Option` `struct` an
`interface` instead, which will allow differentiation of `Option`s, and
move a lot of the logic identifying types of options out of the parsing
logic and into formalized types that confirm to the `Option` interface.

Most of the line changes are just moving code or renaming.