Upgrade testbazel to 8.1.1 with bzlmod (#9028)
We have a `testbazel` utility for use in bazel integration tests, which
provides a `bazel` launcher that points to a pre-extracted
`--install_base` dir from runfiles. This setup is now outdated - it was
using bazel version 5.x and was using deprecated WORKSPACE files. This
PR upgrades us to 8.x with bzlmod (MODULE.bazel + MODULE.bazel.lock).

I noticed that as a result of this upgrade, bazel started making network
requests to bcr during the test, and some builds started failing since
they aren't networking-enabled when run on RBE (e.g. `remote_asset_test`
and `remote_cache_test` were not setting `dockerNetwork=bridge` in
exec_properties, which means they were not using the network).

This PR makes a few changes to make sure we continue to avoid network
requests in these existing tests, as well as remove network requirements
from some other tests which were using the network:
- The rule which pre-extracts bazel now also has to do a `bazel fetch`
to pre-warm the repository_cache, and also now exports a
MODULE.bazel.lock file. I had some trouble with this part, so it's maybe
worth an extra careful look :sweat_smile: . Just running a plain `bazel
fetch` (with no args) seems to download a ton of dependencies for some
reason, but running `bazel fetch //...` on some dummy genrules / shell
rules works as expected.
- The launcher also now also attaches a `bazelrc` to each build which
sets `--repository_cache` to the pre-warmed repository cache. This adds
a little bit of complexity to the launcher unfortunately, but I think
it's a good tradeoff to avoid making network requests which should both
reduce flakiness and improve test speed.
- We now have a utility function `InitWorkspace` which makes sure
MODULE.bazel.lock is added to the workspace, which also seems to be
necessary to avoid requests. The launcher also checks whether
MODULE.bazel.lock exists and if not, will fail the build saying that
`InitWorkspace` must be called.