Always retry if CloseAndRecv fails (#9045)
When `ExecuteTaskAndStreamResults` fails, we already have logic to
unconditionally retry if `PublishOperationDone` fails:
https://github.com/buildbuddy-io/buildbuddy/blob/7d5e90131e7fe59d90ee1b9819bab9398d9ece00/enterprise/server/remote_execution/executor/executor.go#L236-L252

The reasoning behind this PR is:
* `PublishOperationDone` _may or may not_ surface a streaming error (if
there is one), since `Send` on gRPC streams is asynchronous.
* `CloseAndRecv` will _always_ surface a streaming error (if there is
one) since it blocks until either there is an error or the server sends
the final reply on the stream.
* We were OK with unconditionally retrying if `PublishOperationDone`
returned a stream error.
* Therefore, we should also be OK with unconditionally retrying if
`CloseAndRecv` fails as well.

Note: in the future, it would be nice to get rid of PublishOperation
entirely and have the task leaser handle status updates as part of the
TaskLease stream. The retry logic is getting confusing, and the
unconditional retry is not ideal, especially if the client is bazel and
is already capable of retrying on its own.