Don't show an error banner after successfully logging out (#8140)
When clicking "Logout" we show an error banner to the user. This is a
tad confusing because we logged out successfully. It also means that we
have to opt out some of our webdriver tests (which test the logout
behavior) from a [new
assertion](https://github.com/buildbuddy-io/buildbuddy/pull/8128) that
makes sure we don't display error banners during tests.

This PR fixes the logout handler to stop returning an error, so that we
will no longer show the error banner. Instead, we'll just redirect to
the login page. I considered showing a "success" banner in this case but
I don't think it's really necessary since it's pretty clear the user is
logged out, as they're now seeing the login page.