executions: allow double-reading from clickhouse (#8900)
Behind a flag, allow reading execution data from both clickhouse and
mysql. The data that is read includes the buffered rows in redis that
are waiting to be flushed to clickhouse as part of a batch.

This is a necessary step for the MySQL => ClickHouse migration, but also
has the immediate benefit of letting us view additional execution
metadata in the UI that is only available in ClickHouse. In particular,
executor_hostname is only available in ClickHouse, but it would be super
useful to be able to see the executor hostnames in the executions
listing and to be able to drilldown by executor hostname (assuming you
either self-host, or are a member of the shared-executors group). Target
label, action mnemonic, and output_path are also only available in
ClickHouse, so we could update the executions page to show those instead
of / in addition to the command snippet.

This PR also adds an E2E webdriver test which has this new flag enabled.
It does a controlled execution and walks it through the progress stages,
and makes sure we can see the executions in the browser. This test
uncovered various bugs in how we're handling certain DB errors, so this
PR also fixes those.