Improve signal handling/logging in healthchecker (#8990)
- Switch from `fmt.Println()` to `log.Infof()` so that shutdown logs are
handled consistently, using structured logging if enabled. There is a
comment about the log output being wonky due to being called in a signal
handler, but using `log.Infof` seems to work just fine when I tested
locally, and don't see why the log output would be wonky since the
signal handler is just a regular goroutine, but lmk if I'm missing
something.
- When running in a terminal, print a newline when receiving a signal to
work around the `^C` echoed to the terminal from causing extra
indentation and making the logs harder to read. (I suspect the "wonky"
comment might have been referring to this echoing behavior, but the same
problem happens with `log.Infof` so I'm not sure.)
- If the user presses Ctrl+C several times repeatedly, add more feedback
to the logs that we're still shutting down. If the user pressed Ctrl+C
more than 3 times, force-quit (I've seen this behavior in a few other
apps, e.g. bazel and docker-compose). I often find myself mashing Ctrl+C
and it (frustratingly) doesn't work, so I have to open a separate
terminal and SIGKILL the app and the executor. One example of this is
when running a single app and executor locally, and trying to shut down
while an execution is still in progress. The app and executor get
deadlocked because they are both trying really really hard to make sure
the task gets re-enqueued, so they ignore the context cancellation that
happens when the gRPC server is hard-stopped, and become unresponsive to
Ctrl+C. Another example is when trying to use the health checker inside
a bazel test, since we install the signal handlers even in tests.
(Arguably, both of these examples are their own issues that can be
addressed separately, but there are other ways I've seen the server get
stuck especially while developing features locally - e.g. writing an
infinite loop or receiving from a channel without also receiving from
`ctx.Done()`. If the user is mashing Ctrl+C it just seems generally
sensible to force-kill the app.)
- Add logging of the specific signals that we're receiving, instead of
always logging "Caught interrupt signal" (this is inaccurate/confusing
if we actually got SIGTERM)