Compress uploads more consistently (#8371)
We used to (configurably) apply upload compression only to `UploadFile`,
not e.g. `UploadBlob`. Make this more principled by enabling compression
on more code paths, but only if the file size is large enough to be
worth it (see the linked Bazel discussion).

Always enable compression for the remote asset API as files fetched from
the internet are overwhelmingly likely to be larger than 100 bytes
anyway.