Fix a goroutine leak in pubsub.Subscribe
I found this while trying to figure out why some tasks have way more goroutines than others. I don't think this is the cause, and it's no that severe, but might as well fix it.

Right now, if the context expires while the monitoring goroutine is calling `checkMonitoredChannelExists`, then `msgChan` will get closed, and the third (merge) goroutine will exit, so nothing will be listening t `monChan`.

If fixed this by putting each error send in a select with a `<-ctx.Done()`.