[oci registry] do not rely on Docker-Content-Digest header for caching layers (#9067)
Docker Hub does not (or does not always) return a Docker-Content-Digest
header for layers.
That's fine, you have to know a layer's digest to even request it.
However, writing layers to the cache depend on the Docker-Content-Digest
header appearing in the response from the upstream registry.

This change makes sure the test registry does not return
Docker-Content-Digest for blobs (layers), and patches the caching code
to not rely on this header.