[raft] post commit hook (#9085)
Right now, we execute a post commit hook twice:
- at the end of a prepare transaction
- inside CommitTransaction,

add a transaction hook message in transaction statement which includes
the
hook and also an enum to indicate when the hook should be executed

The current hook snapshotCluster should be executed when the txn is
committed,
but I'm going to introduce a new hook in a following pr where it should
be
executed at the end of the prepared stage.

The txn coordinator is reponsible for setting the hook in the batch
request.
This way we can remove the handleHook in the CommitTransaction step.