Simplify and improve ociruntime cgroup setup (#9113)
- Recursively enable ancestor cgroup controllers if they are not already
enabled. `crun` effectively does this already, but only after our first
failed attempt to enable the controllers. This results in warnings being
printed to the logs on executor startup (when it runs the warmup
container). Fixes
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4842
- Use `--keep` to prevent the runtime from immediately deleting the
container cgroup when the task completes. This removes the need for the
parent/child setup and also is a prerequisite for being able to detect
when the `pids.max` limit exceeded, since pid events are only reported
to the child cgroup which currently is being cleaned up before we have
the chance to read from it. This should also let us get more accurate
resource usage measurements (in a later PR), since we can now take one
final measurement just after the task completes. Today, we're using
polling, so we're potentially losing ~50ms of data or more, depending on
load.