Fix action merging handling of `do_not_cache` and `skip_cache_lookup` (#8746)
`skip_cache_lookup` should not prevent action merging (it currently
does), but `do_not_cache` should (it currently doesn't). The former is
used by Bazel when an action result is discovered to be invalid, but
that can't have happened while the merged execution is still executing.
The latter powers `--runs_per_test` and thus won't be useful to detect
flakes if action merging is used.

Since `--runs_per_test` sets both fields to `true`, we did correctly
disable action merging for it, but for the wrong reason. Other code
paths may not do this though.

See
https://github.com/bazelbuild/remote-apis/commit/cfe8e540cbb424e3ebc649ddcbc91190f70e23a6
for the intended semantics.