Log a warning if pids.max is hit (#9138)
When the cgroup2 file `pids.events` contains a line like `max N` (where
`N` is a number >0), this means the container tried to create a process
but failed due to the `pids.max` limit being enforced.

With this change, we'll log a warning when this happens, so that we can
diagnose these issues better. These issues are unlikely now that #9098
is merged, but if these do pop up in the future we'll at least be able
to see them in the logs.

Note: this change depends on
https://github.com/buildbuddy-io/buildbuddy/pull/9113, because
`pids.events` is only reported to the container cgroup, and is not
bubbled up to any parent cgroups. Prior to that change, `crun` would
immediately delete the container cgroup once the task exited, preventing
us from detecting these events.