Refactor ByteStreamServer.Write() to make it easier to use directly in ByteStreamServerProxy (#9105)
The idea here is to move as much of the business logic of handling
ByteStream writes as possible into the `WriteState` struct in
`byte_stream_server.go`, making the actual `Write()` function only
responsible for pulling frames off of the write stream, and sending them
to the `WriteState`. This way, the `WriteState` can be used directly in
`byte_stream_server_proxy` without loss of functionality. See
https://github.com/buildbuddy-io/buildbuddy/pull/9106 for how this will
be used.

The alternative to this is to do the splitting at the gRPC level, but
that ends up being a little more complicated IMO due to the async nature
of the streams.

Related issues:
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4782