Introduce `newUploadWriteCloser`, use it in `uploadFromReader` (#9097)
In the OCI registry, I eventually want to have a `Writer` interface that
writes to the CAS (so I can write to both the response and the CAS).
This change introduces an `uploadWriteCloser` that uploads to the CAS.
If anyone calls `Upload`, `UploadFromReader`, or `UploadProto` with
`IDENTITY` compression, they will be using the `uploadWriteCloser` under
the hood.