[raft] haveLease should return false when it shoudn't have lease (#9156)
This can cause test-flakiness.

In tests, we often want to find the store with the range leases. There
are small
windows of time where two stores both hold leases, with one of them
should not
have it hold. In these cases, if we grab the store which should hold the
lease,
it can cause the test to fail.