Use `arg.Append` to append to cli bazel args (#9054)
Switch to using `arg.Append` instead of `append(args, ...)` to append to
bazel args. This helps protect us from cases where there may be a `--`
in the args we are appending to, and also abstracts the call so we can
use real parsing in the future to do appending.

Also enforces `JoinExecutableArgs` not adding an additional `--` if one
already exists.