Reinstate "[SS] Skip accessing the remote cache when proxy is enabled" (#9072) (#9108)
This reverts commit fd32eb407b14cdeb56dad1b004965b52ccb51958.

Last week, I noticed in the metrics we had some SkipRemote requests
recorded in the cacheproxy. I only expected to see SkipRemote requests
coming from workflows, which are currently not running with the proxy,
so I reverted the PR out of caution.

After looking into it more, I think those are coming from requests for
the snapshot version. Firecracker RBE actions only save snapshots
locally, but they still request version data from the remote cache.

The first commit in this PR is the change that was reverted. In the
second commit, I remove SkipRemote from requests for the snapshot
version, so we'll always save that in the remote cache. This will
guarantee that any version invalidations are applied universally