Test that zstd compression and decompression is lossless (#9183)
This change adds a test that takes all the publicly available zstd
compressors (`CompressZstd` and `NewZstdCompressingReader`) and zstd
decompressors (`DecompressZstd`, `NewZstdDecompressor`,
`NewZstdDecompressingReader`) and makes sure that compressing and
decompressing through each pair produces the same string.

This change also removes `Test_NewZstdDecompressingReader`, as
`TestLossless` is a superset of that test.

I'm about to introduce `NewZstdCompressingWriter` in another change, and
I want to be able to test that it too is lossless.