[raft] Add staging replica descriptors in range descriptor (#9189)
https://github.com/buildbuddy-io/buildbuddy-internal/issues/4910

Added a staging field in range descriptor.

When we add a replica, we will 
1. add to the staging field in range descriptor
2. start the shard as a non-voter, catch up, promote to voter
3. move the replica descriptor from staging to regular replica
descriptors

In the sender, when we try replicas, we will also try staging replicas,
so that
we won't ended up fail the writes with no lease found for replicas when
the
leader is moved to the newly added replica before the range descriptor
is
updated.

Also, fixed some issues with AddNodeToCluster tests:
- disable driver
- add s3 to the candidate stores when we try to get the store with range
lease.

The AddNodeToCluster tested succeeded with 2 500 runs:

https://app.buildbuddy.io/invocation/b5d1949d-63ab-42f1-b772-b7323065b41f

https://app.buildbuddy.io/invocation/0a847409-42c3-41e1-b7e0-8e23d401c072

Note: this change doesn't change zombie janitor yet. With the staging
replica descriptor, we can remove the janitor to check for non-voters,
and instead let zombie janitor check staging replica descriptor. Also,
zombie janitor needs to clean up staging replica after it removes zombie
shards. I will accomplish these in a future PR.