vfs: refresh attrs in Lookup if needed (#9190)
I don't know how to write a simple test for this. There doesn't seem to
be a way for user space code to exercise Lookup w/o also calling
GetAttr.

overlayfs is likely able to do it because it's running in the kernel and
is able to use lower-level APIs. Perhaps the only option is to add an
overlayfs test to vfs_test?