Override the user_id and group_id used for experiments (#9204)
The executor sends a JWT in the gRPC metadata using its API key, so we
can't use the group ID from that for experiments. Instead, use the JWT
from the request to override the group ID (and user ID).

This is a short term workaround while I figure out how to get the task
JWT in the context, and if we should do that.