Deflake executions_clickhouse_test (#9223)
- Check bazel status in `defer` instead of `t.Cleanup()`, since the ctx
will be canceled just before cleanup funcs are run. If the bazel command
hasn't yet finished, the ctx cancellation kills the bazel invocation and
makes it return a `signal: killed` error, which fails the test
unnecessarily.
- Increase webdriver implicit wait timeout. The current 1s timeout is
causing too much self-inflicted pain, given that tests can be throttled
when the system is under high load.
- Fix "stale element reference" error in `waitForExecutionsToAppear`.
- Increase `EstimatedComputeUnits` to carve out more space for the test
and reduce the risk of contention-related failures.