[raft] fix TestDownReplicate (#9215)
This PR https://github.com/buildbuddy-io/buildbuddy/pull/8941 changed
the driver
so that remove replica as an raft cluster member and remove data becomes
two
driver action. This introduced the flakiness in the test; because some
times,
when we verify whether the data is removed, the remove data action has
not been
run.

The test passes with --runs_per_test=500 on top of the change to
decrease raft disk usage in test:

https://app.buildbuddy.io/invocation/1ab84174-781d-4709-927d-83f9ff233d7b

https://app.buildbuddy.io/invocation/fc3d6700-8f4d-4530-aedb-a133184069ae

https://github.com/buildbuddy-io/buildbuddy-internal/issues/4192