Roll-forward: Make ByteStreamServerProxy call ByteStreamServer directly (#9213)
This PR rolls forward
https://github.com/,buildbuddy-io/buildbuddy/pull/9106 and
https://github.com/buildbuddy-io/buildbuddy/pull/9169 which, together
were reverted in https://github.com/buildbuddy-io/buildbuddy/pull/9208
and fixes a case where we mistakenly propagated EOF errors from the
remote to the caller.

Specifically, the bug was that this previous logic:
https://github.com/buildbuddy-io/buildbuddy/blob/afcfc41037a4daccb031d576a013234fd2a3f186/enterprise/server/byte_stream_server_proxy/byte_stream_server_proxy.go#L324-L332
would handle EOF errors from the remote write stream by causing a
`SendAndClose()` call on the client stream. This was broken in the two
PRs I mentioned at the top, here:
https://github.com/buildbuddy-io/buildbuddy/blob/90173e817f5c63b3da1bf296c3a5263fb3b849a2/enterprise/server/byte_stream_server_proxy/byte_stream_server_proxy.go#L279-L281

The last commit in this PR fixes the bug by handling EOF there in the
same way it was previously handled.