Don't send API key to LeaseTask (#9225)
Since https://github.com/buildbuddy-io/buildbuddy/pull/9125, we send the
task's JWT to LeaseTask. This allows experiments to trigger on the
task's group ID.

When also sending the API key, it takes precedence, so we can't get the
task's group ID. Since we don't need the API key any more, stop sending
it.