[SS] Always look for a local manifest before remote manifest (#9233)
In an effort to re-push #9221 , break out some of the changes to make
them less risky and easier to review.

With #9221, we will write fewer remote snapshots and should try to use a
local snapshot instead. This PR should have no effect until #9221 is
merged, because workflows won't write local manifests so we'll always
fall back to reading the remote manifest