[SS] Support starting from a local manifest but fetching chunks remotely (#9232)
I had to rollback #9084 because it made an assumption that if a snapshot
was cached locally, all the chunks existed locally as well. This
assumption does not hold true, because we lazily load snapshot chunks.
So the following situation is possible:

1. [Executor A] Write snapshot chunks { 0 1 2 3 4 } to the remote cache
2. [Executor B] Fetch the remote snapshot. Only lazily load { 2 4 }. Now
only { 2 4} will be cached locally, but the local manifest will still
reference all chunks { 0 1 2 3 4}
3. [Executor B] Try to start from the local snapshot from # 2. We fail
to use the local snapshot, because `checkAllArtifactsExist` reports that
we don't have chunks {0 1 3} in the local filecache.

After this PR, in # 3 we will allow you to start from the local
snapshot, fetching {2 4} from the local cache and {0 1 3} from the
remote cache. This fallback behavior is already supported in
`copy_on_write::fetch`