Fix crash due to thread-unsafe usage of tokenizers in codesearch
Tokenizers are now created on-demand and cached on Writer instances.
This avoids reuse of a single tokenizer instance across multiple Writers.

Note that there is a larger issue here around updating the same document
concurrently with multiple writers, but this change is intended to get us
back to a non-crashing state before addressing this larger issue.