Set worker enqueued timestamp in auxiliary metadata (#9267)
This will let us track the timestamp at which the task was added to the
executor's local queue. This can be significantly different from the
queued_timestamp computed in the apps, specifically in the case where
the execution was interrupted and re-enqueued due to retries.

The plan with this new field is to render the worker enqueue timestamp
in the trace viewer on the execution page, which can help clarify this
retry behavior. This also paves the way for reporting "queue wait time"
metrics for each executor. One hypothesis is that long queue wait times
might be a useful signal that might help improve our autoscaling
behavior. Having these metrics would help us understand how executor
pools would behave if we incorporated these metrics.