Fix crash due to thread-unsafe usage of tokenizers in codesearch (#9277)
Tokenizers are not threadsafe. In this change, Tokenizers are now
created on-demand and cached on Writer instances. This avoids reuse of a
single tokenizer instance across multiple Writers. Writers themselves
are not shared across threads currently. This crash was introduced in
#9255.

Note that there is a larger issue here around updating the same document
concurrently with multiple writers, but this change is intended to get
us back to a non-crashing state before addressing this larger issue.