[oci registry] separate functions that read from or write to AC or CAS (#9280)
This change is a mechanical change to separate all functions that
interact with the AC and CAS from the rest of the OCI registry mirror
code.

A later change will move these cache functions to `oci.go` and make them
public.
That later change will then create a dependency from `ociregistry` on
`oci`.

This is all setup for caching (potentially) private OCI images in
`oci.Resolver.Resolve(...)`.