Bugfix: Prevent codesearch from returning matches that don't fully match (#9282)
This change fixes a bug in codesearch where, in certain circumstances,
docs that shouldn't match were returned. This happened most often when
short tokens were included in the query.

There are 2 relevant phases in a query:
1. Find docs that match according to the n-gram index
2. Post filter and rank (score) these documents

In phase 1, we sometimes over-retrieve documents, because n-gram indices
are incapable of filtering on tokens < 3 characters long. For example,
searching for "foo ba" will return all documents that have "foo", but
will not filter at all on "ba". This also shows up in cases like
"foo.ba", where `.` is treated as single-char wildcard. In these cases,
post-filtering is needed to trim out all the non-matches.

In phase 2, we score all the documents. Prior to this change, we didn't
filter any documents out during this step (aside from honoring
limit/offset query parameters).

In this change, we now drop documents that have a score 0. A score of 0
indicates that the resulting document didn't actually match the full
query. So a document with "foo" but not "ba" would receive a score of 0.

One additional related change is that we used to add an additional
matcher on filename for content searches. The ultimate intention here
was to match on filename or content when not using a limiter in the
query (e.g. "foo" matches files with "foo" in them, but also files with
"foo" in the filename). However, this extra matcher didn't cause
filenames to match, because the squery didn't include the relevant
clauses to retrieve them from the index in the first place. We can
revisit supporting this functionality in the future, but that would
require a larger change, and also requires a more complex approach to
scoring (scoring would need to support disjunctions).

Note that this bug wasn't very visible, because many cases were papered
over by the fact that the highlighter filtered out docs with no matching
regions. It was visible with filename matches though.