[oci registry] Move OCI image caching functions to `ocicache` package (#9289)
This change is a mechanical change to move all functions that interact
with the AC or CAS from the `ociregistry` package to the new `ocicache`
package.
Since the `ociregistry` is currently the only caller, its test suite
also provides coverage for these cache functions.
Testing the cache functions directly will come in a later change.