Add execution collector methods needed to replace MySQL (#9258)
Split out from #9211 but added tests and cleaned things up slightly.

Currently, only completed executions are stored in Redis, and only while
they are waiting for the invocation to be completed.

This PR adds the functionality needed to store in-progress[^1]
executions as well, which would allow us to replace MySQL reads/writes
with redis reads/writes. Once each execution is completed and moved to
the existing list of completed executions (for each invocation), we'll
delete the in-progress execution state in favor of the per-invocation
execution lists (or the ClickHouse executions for the invocation, if the
invocation is complete).

The new functionality is not used anywhere yet.

Context:
https://docs.google.com/document/d/1qf-veFLjAVbXMCtAn1a5IVymWIAaqOBpoQiUO5EOPyk/view

[^1]: One potentially confusing thing is that we write the COMPLETED
execution state to this entry in redis, even though it's referred to as
the "in-progress" execution state. I wanted some way to differentiate
this new list from the per-invocation execution lists which _also_
contain completed executions, but are just waiting for the invocation to
complete (so that we can fill out the invocation-specific fields before
flushing to ClickHouse). I figured since we move the execution from this
"in-progress" list to the per-invocation "waiting to be flushed" lists
immediately after it's complete, the "in-progress" naming almost always
reflects the actual data that we're storing there, so this naming should
hopefully be intuitive enough, but I'm open to better suggestions here
:)